requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
aiosqlite>=0.19.0
anthropic>=0.40.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
psutil>=5.9.0
pydantic-settings>=2.1.0
